fix: integrate all audit execution in run audit command
