+++ src/support/slack/commands/run-audit.js
@@ -10,9 +10,15 @@
  * governing permissions and limitations under the License.
  */
 
+import { isNonEmptyArray, isNonEmptyObject, isValidUrl } from '@adobe/spacecat-shared-utils';
 import BaseCommand from './base.js';
 
-import { extractURLFromSlackInput, postErrorMessage } from '../../../utils/slack/base.js';
+import {
+  extractURLFromSlackInput,
+  parseCSV,
+  postErrorMessage,
+  postSiteNotFoundMessage,
+} from '../../../utils/slack/base.js';
 
 import { triggerAuditForSite } from '../../utils.js';
 
@@ -30,13 +36,71 @@ function RunAuditCommand(context) {
   const baseCommand = BaseCommand({
     id: 'run-audit',
     name: 'Run Audit',
-    description: 'Run audit for a previously added site. Runs lhs-mobile by default if no audit type parameter is provided.',
+    description: 'Run audit for a previously added site. Runs lhs-mobile by default if no audit type parameter is provided. Runs all audits if audit type is `all`',
     phrases: PHRASES,
     usageText: `${PHRASES[0]} {site} [auditType (optional)]`,
   });
 
   const { dataAccess, log } = context;
   const { Configuration, Site } = dataAccess;
+
+  /**
+   * Runs an audit for the given site.
+   * @param {string} baseURL - The base URL of the site.
+   * @param {string} auditType - The type of audit to run.
+   * @param {object} slackContext - The Slack context object.
+   * @returns {Promise} A promise that resolves when the operation is complete.
+   */
+  const runAuditForSite = async (baseURL, auditType, slackContext) => {
+    const { say } = slackContext;
+
+    try {
+      const site = await Site.findByBaseURL(baseURL);
+      const configuration = await Configuration.findLatest();
+
+      if (!isNonEmptyObject(site)) {
+        await postSiteNotFoundMessage(say, baseURL);
+        return;
+      }
+
+      if (auditType === 'all') {
+        const enabledAudits = configuration.getEnabledAuditsForSite(site);
+
+        if (!isNonEmptyArray(enabledAudits)) {
+          await say(`:warning: No audits configured for site \`${baseURL}\``);
+          return;
+        }
+
+        await Promise.all(
+          enabledAudits.map(async (enabledAuditType) => {
+            try {
+              await triggerAuditForSite(site, enabledAuditType, slackContext, context);
+            } catch (error) {
+              log.error(`Error running audit ${enabledAuditType.id} for site ${baseURL}`, error);
+              await postErrorMessage(say, error);
+            }
+          }),
+        );
+        say(':white_check_mark: All audits triggered successfully.');
+      } else {
+        if (!configuration.isHandlerEnabledForSite(auditType, site)) {
+          await say(`:x: Will not audit site '${baseURL}' because audits of type '${auditType}' are disabled for this site.`);
+          return;
+        }
+        await triggerAuditForSite(site, auditType, slackContext, context);
+
+        let message = `:white_check_mark: ${auditType} audit check is triggered for ${baseURL}\n`;
+        if (auditType === LHS_MOBILE) {
+          message += `:adobe-run: In a minute, you can run @spacecat get site ${baseURL}`;
+        }
+
+        await say(message);
+      }
+    } catch (error) {
+      log.error(`Error running audit ${auditType} for site ${baseURL}`, error);
+      await postErrorMessage(say, error);
+    }
+  };
 
   /**
    * Validates input, fetches the site
@@ -48,41 +112,55 @@ function RunAuditCommand(context) {
    * @returns {Promise} A promise that resolves when the operation is complete.
    */
   const handleExecution = async (args, slackContext) => {
-    const { say } = slackContext;
+    const { say, files, botToken } = slackContext;
 
     try {
       const [baseURLInput, auditTypeInput] = args;
 
       const baseURL = extractURLFromSlackInput(baseURLInput);
-
-      if (!baseURL) {
-        await say(':warning: Please provide a valid site url.');
-        return;
-      }
-
-      const site = await Site.findByBaseURL(baseURL);
-
-      if (!site) {
-        await say(`:x: '${baseURL}' was not added previously. You can run '@spacecat add site ${baseURL}`);
-        return;
-      }
-
+      const hasValidBaseURL = isValidUrl(baseURL);
+      const hasFiles = isNonEmptyArray(files);
       const auditType = auditTypeInput || LHS_MOBILE;
-      const configuration = await Configuration.findLatest();
 
-      if (!configuration.isHandlerEnabledForSite(auditType, site)) {
-        await say(`:x: Will not audit site '${baseURL}' because audits of type '${auditType}' are disabled for this site.`);
+      if (!hasValidBaseURL && !hasFiles) {
+        await say(baseCommand.usage());
         return;
       }
 
-      await triggerAuditForSite(site, auditType, slackContext, context);
-
-      let message = `:white_check_mark: ${auditType} audit check is triggered for ${baseURL}\n`;
-      if (auditType === LHS_MOBILE) {
-        message += `:adobe-run: In a minute, you can run @spacecat get site ${baseURL}`;
+      if (hasValidBaseURL && hasFiles) {
+        await say(':warning: Please provide either a baseURL or a CSV file with a list of site URLs.');
+        return;
       }
 
-      await say(message);
+      if (hasFiles) {
+        if (files.length > 1) {
+          await say(':warning: Please provide only one CSV file.');
+          return;
+        }
+
+        const file = files[0];
+        if (!file.name.endsWith('.csv')) {
+          await say(':warning: Please provide a CSV file.');
+          return;
+        }
+
+        const csvData = await parseCSV(file, botToken);
+
+        say(`:adobe-run: Triggering ${auditType} audit for ${csvData.length} sites.`);
+
+        await Promise.all(
+          csvData.map(async (row) => {
+            const [csvBaseURL] = row;
+            if (isValidUrl(csvBaseURL)) {
+              await runAuditForSite(csvBaseURL, auditType, slackContext);
+            } else {
+              await say(`:warning: Invalid URL found in CSV file: ${csvBaseURL}`);
+            }
+          }),
+        );
+      } else if (hasValidBaseURL) {
+        await runAuditForSite(baseURL, auditType, slackContext);
+      }
     } catch (error) {
       log.error(error);
       await postErrorMessage(say, error);
