+++ test/support/slack/commands/run-audit.test.js
@@ -15,6 +15,7 @@
 import { expect, use } from 'chai';
 import sinonChai from 'sinon-chai';
 import sinon from 'sinon';
+import nock from 'nock';
 
 import RunAuditCommand from '../../../../src/support/slack/commands/run-audit.js';
 
@@ -48,7 +49,7 @@ describe('RunAuditCommand', () => {
       const command = RunAuditCommand(context);
       expect(command.id).to.equal('run-audit');
       expect(command.name).to.equal('Run Audit');
-      expect(command.description).to.equal('Run audit for a previously added site. Runs lhs-mobile by default if no audit type parameter is provided.');
+      expect(command.description).to.equal('Run audit for a previously added site. Runs lhs-mobile by default if no audit type parameter is provided. Runs all audits if audit type is `all`');
     });
   });
 
@@ -89,9 +90,9 @@ describe('RunAuditCommand', () => {
     it('responds with a warning for an invalid site url', async () => {
       const command = RunAuditCommand(context);
 
-      await command.handleExecution([''], slackContext);
+      await command.handleExecution(['invalid-url'], slackContext);
 
-      expect(slackContext.say.calledWith(':warning: Please provide a valid site url.')).to.be.true;
+      expect(slackContext.say.calledWith(command.usage())).to.be.true;
     });
 
     it('informs user if the site was not added previously', async () => {
@@ -100,7 +101,7 @@ describe('RunAuditCommand', () => {
 
       await command.handleExecution(['unknownsite.com'], slackContext);
 
-      expect(slackContext.say.calledWith(':x: \'https://unknownsite.com\' was not added previously. You can run \'@spacecat add site https://unknownsite.com')).to.be.true;
+      expect(slackContext.say.calledWith(':x: No site found with base URL \'https://unknownsite.com\'.')).to.be.true;
     });
 
     it('informs user when error occurs', async () => {
@@ -110,6 +111,162 @@ describe('RunAuditCommand', () => {
       await command.handleExecution(['some-site.com'], slackContext);
 
       expect(slackContext.say.calledWith(':nuclear-warning: Oops! Something went wrong: Test Error')).to.be.true;
+    });
+
+    it('trigger all audits for a valid site', async () => {
+      dataAccessStub.Site.findByBaseURL.resolves({
+        getId: () => '123',
+      });
+      dataAccessStub.Configuration.findLatest.resolves({
+        getEnabledAuditsForSite: () => ['lhs-mobile', 'lhs-desktop'],
+      });
+
+      const command = RunAuditCommand(context);
+      await command.handleExecution(['validsite.com', 'all'], slackContext);
+
+      expect(slackContext.say.called).to.be.true;
+      expect(slackContext.say.firstCall.args[0]).to.equal(':white_check_mark: All audits triggered successfully.');
+      expect(sqsStub.sendMessage.called).to.be.true;
+    });
+
+    it('triggers all audits for all sites specified in a CSV file', async () => {
+      dataAccessStub.Site.findByBaseURL.resolves({
+        getId: () => '123',
+      });
+      dataAccessStub.Configuration.findLatest.resolves({
+        getEnabledAuditsForSite: () => ['lhs-mobile', 'lhs-desktop'],
+      });
+      const fileUrl = 'https://example.com/sites.csv';
+      slackContext.files = [
+        {
+          name: 'sites.csv',
+          url_private: fileUrl,
+        },
+      ];
+      nock(fileUrl)
+        .get('')
+        .reply(200, 'https://site.com,uuidv4\n'
+          + 'https://valid.url,uuidv4');
+
+      const command = RunAuditCommand(context);
+      await command.handleExecution(['', 'all'], slackContext);
+
+      expect(slackContext.say.called).to.be.true;
+      expect(slackContext.say.firstCall.args[0]).to.equal(':adobe-run: Triggering all audit for 2 sites.');
+      expect(slackContext.say.secondCall.args[0]).to.equal(':white_check_mark: All audits triggered successfully.');
+      expect(sqsStub.sendMessage.called).to.be.true;
+    });
+
+    it('handles both site URL and CSV file', async () => {
+      const command = RunAuditCommand(context);
+      slackContext.files = [
+        {
+          name: 'sites.csv',
+          url_private: 'https://example.com/sites.csv',
+        },
+      ];
+      await command.handleExecution(['site.com'], slackContext);
+      expect(slackContext.say.calledWith(':warning: Please provide either a baseURL or a CSV file with a list of site URLs.')).to.be.true;
+    });
+
+    it('handles multiple CSV files', async () => {
+      const command = RunAuditCommand(context);
+      slackContext.files = [
+        {
+          name: 'sites1.csv',
+          url_private: 'https://example.com/sites1.csv',
+        },
+        {
+          name: 'sites2.csv',
+          url_private: 'https://example.com/sites2.csv',
+        },
+      ];
+      await command.handleExecution(['', 'all'], slackContext);
+      expect(slackContext.say.calledWith(':warning: Please provide only one CSV file.')).to.be.true;
+    });
+
+    it('handles non-CSV file', async () => {
+      const command = RunAuditCommand(context);
+      slackContext.files = [
+        {
+          name: 'sites.txt',
+          url_private: 'https://example.com/sites.txt',
+        },
+      ];
+      await command.handleExecution(['', 'all'], slackContext);
+      expect(slackContext.say.calledWith(':warning: Please provide a CSV file.')).to.be.true;
+    });
+
+    it('handles CSV file with no data', async () => {
+      const command = RunAuditCommand(context);
+      slackContext.files = [
+        {
+          name: 'sites.csv',
+          url_private: 'https://example.com/sites.csv',
+        },
+      ];
+      nock('https://example.com')
+        .get('/sites.csv')
+        .reply(200, 'invalid-url,uuidv4\n');
+
+      await command.handleExecution(['', 'all'], slackContext);
+      expect(slackContext.say.calledWith(':warning: Invalid URL found in CSV file: invalid-url')).to.be.true;
+    });
+
+    it('handles site with no enable audits', async () => {
+      dataAccessStub.Site.findByBaseURL.resolves({
+        getId: () => '123',
+      });
+      dataAccessStub.Configuration.findLatest.resolves({
+        getEnabledAuditsForSite: () => [],
+      });
+
+      const command = RunAuditCommand(context);
+      await command.handleExecution(['validsite.com', 'all'], slackContext);
+
+      expect(slackContext.say.called).to.be.true;
+      expect(slackContext.say.firstCall.args[0]).to.equal(':warning: No audits configured for site `https://validsite.com`');
+    });
+
+    it('handles error while triggering audits', async () => {
+      const errorMessage = 'Failed to send message';
+      dataAccessStub.Site.findByBaseURL.resolves({
+        getId: () => '123',
+      });
+      dataAccessStub.Configuration.findLatest.resolves({
+        getEnabledAuditsForSite: () => ['lhs-mobile', 'lhs-desktop'],
+      });
+      sqsStub.sendMessage.rejects(new Error(errorMessage));
+
+      const command = RunAuditCommand(context);
+      await command.handleExecution(['validsite.com', 'all'], slackContext);
+
+      expect(slackContext.say.called).to.be.true;
+      expect(slackContext.say.firstCall.args[0]).to.equal(`:nuclear-warning: Oops! Something went wrong: ${errorMessage}`);
+    });
+
+    it('handles error when site cannot be found', async () => {
+      const errorMessage = 'Invalid site URL';
+      dataAccessStub.Site.findByBaseURL.rejects(new Error(errorMessage));
+      const command = RunAuditCommand(context);
+      await command.handleExecution(['invalidsite.com', 'all'], slackContext);
+      expect(slackContext.say.calledWith(`:nuclear-warning: Oops! Something went wrong: ${errorMessage}`)).to.be.true;
+    });
+
+    it('handles error when obtaining CSV failed', async () => {
+      const command = RunAuditCommand(context);
+      slackContext.files = [
+        {
+          name: 'sites.csv',
+          url_private: 'https://example.com/sites.csv',
+        },
+      ];
+      nock('https://example.com')
+        .get('/sites.csv')
+        . reply(401, 'Unauthorized');
+
+      await command.handleExecution(['', 'all'], slackContext);
+      expect(slackContext.say.calledWith(':nuclear-warning: Oops! Something went wrong: CSV processing failed: Authentication failed: Invalid Slack token.')).to.be.true;
     });
   });
 });
